tests/SampleApps/python/mysql-mysqlconnector-sample/requirements.txt
Flask==1.0.2
mysql-connector==2.1.6